Fix town borders display (#28)
* Fix town borders display

* Remove some unnecessary casts

+++ ja2lib/Strategic/MapScreenInterfaceTownMineInfo.c
@@ -339,8 +339,8 @@ void AddTextToTownBox(void) {
     // Associated Mine: Sector
     swprintf(wString, ARR_SIZE(wString), L"%s:", pwTownInfoStrings[4]);
     AddMonoString(&hStringHandle, wString);
-    GetShortSectorString((INT16)(SectorID16_X(sMineSector)), (INT16)(SectorID16_Y(sMineSector)),
-                         wString, ARR_SIZE(wString));
+    GetShortSectorString(SectorID16_X(sMineSector), SectorID16_Y(sMineSector), wString,
+                         ARR_SIZE(wString));
     AddSecondColumnMonoString(&hStringHandle, wString);
   }
 }
